feat(Project): 3D Added
The game is now entirely functional in 3D.
feat(UI): Improved UI

Added semi permanent UI. Some of it could be improved, such as the main menu, but most of it is good enough.
fix: Removed Videos
feat(Project): Added Polish

This commit added a lot of polish to the game and UI. More could still be used, but this commit especially focused on the juice of the UI.
chore: Added LFS for .MP4s
perf(Player Model): Lessened the Number of Vertices on Model

Lessened the number of vertices on the player model and fixed TweenUI script on death.
build: Build

Build of the game in 3D.